refactor : 자동차의 이름 검증을 Name 객체 안에서 진행, 명확한 변수명으로 수정
